Test case added for validation
